Testes unitarios + delete

+++ src/main/java/br/com/pessoa/servicos/PessoaServico.java
@@ -1,5 +1,6 @@
 package br.com.pessoa.servicos;
 
+import br.com.pessoa.exceptions.PessoaException;
 import br.com.pessoa.repositorios.PessoaRepositorio;
 import br.com.pessoa.repositorios.entidades.Pessoa;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,22 +18,29 @@ public class PessoaServico {
         this.pessoaRepositorio = pessoaRepositorio;
     }
 
-    public void save(Pessoa pessoa){
+    public Pessoa save(Pessoa pessoa){
         if(validarCastro(pessoa))
-            pessoaRepositorio.save(pessoa);
+            return pessoaRepositorio.save(pessoa);
+        return null;
     }
 
     private boolean validarCastro(Pessoa pessoa){
-        return Util.verificarCampoNull(pessoa.getNome())
-                && Util.verificaAltura(pessoa.getAltura())
-                && Util.verificaIdade(pessoa.getIdade());
+        return Util.verificarCampoNaoNull(pessoa.getNome())
+                && Util.verificaAlturaPreenchida(pessoa.getAltura())
+                && Util.verificaIdadePreenchida(pessoa.getIdade());
     }
 
-    public List<Pessoa> listarTodos(){
-        return pessoaRepositorio.findAll();
+    public List<Pessoa> listarTodos() throws PessoaException {
+        List<Pessoa> pessoas = pessoaRepositorio.findAll();
+        if(pessoas == null
+            || pessoas.isEmpty())
+            throw new PessoaException("Nenhuma pessoa encontrada");
+        return pessoas;
     }
 
-    public void deletaPessoa(Long idPessoa){
+    public void deletaPessoa(Long idPessoa) throws PessoaException{
+        if(idPessoa == null || idPessoa <= 0)
+            throw new PessoaException("Não exite pessoa a ser deletada com este id");
         pessoaRepositorio.deleteById(idPessoa);
     }
 }
